src/routes/quizzes/{id}/get.ts
import ManualAssignment from "@models/manual/assignment";
import Manual from "@models/manual/manual";
import Quiz from "@models/quiz/quiz";
import User from "@models/user/user";
import UserRole from "@models/user/user_role";
import { Request, Response } from "express";

const getController = async (req: Request, res: Response): Promise<void> => {
    const {
        dbConnection,
        session: { user_id, current_business_id },
        params: { id },
    } = req;

    const [isAdmin, isManager] = await Promise.all([
        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
        User.isAdmin(dbConnection, current_business_id!, user_id!),
        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
        User.isManager(dbConnection, current_business_id!, user_id!),
    ]);

    let query = dbConnection
        .createQueryBuilder()
        .select(
            "q.id, q.title, q.published, q.prevent_delete, q.prevent_edit, q.max_attempts, q.manual_id"
        )
        .from(Quiz, "q")
        .leftJoin(Manual, "m", "m.id = q.manual_id")
        .leftJoin(ManualAssignment, "ma", "ma.manual_id = m.id")
        .leftJoin(UserRole, "ur", "ur.role_id = ma.role_id")
        .where("m.business_id = :current_business_id", { current_business_id })
        .andWhere("q.id = :id", { id });

    if (!(isAdmin || isManager)) {
        const isAssigned = await query
            .andWhere("ur.user_id = :user_id", { user_id })
            .getRawOne();

        if (!isAssigned) {
            res.sendStatus(403);
            return;
        }

        query = query
            .andWhere("ur.user_id = :user_id", { user_id })
            .andWhere("q.published = :qPublished", {
                qPublished: true,
            })
            .andWhere("m.published = :mPublished", { mPublished: true });
    }

    const quiz = await query.getRawOne();

    res.status(200).send(
        quiz
            ? {
                  ...quiz,
                  prevent_edit: quiz.prevent_edit === 1,
                  prevent_delete: quiz.prevent_delete === 1,
                  published: quiz.published === 1,
              }
            : undefined
    );
};

export default getController;
